requirements.txt
curl_cffi==0.6.2
beautifulsoup4==4.12.2
pymongo==4.6.1
flask==3.0.0
flask-cors==4.0.0
streamlit==1.29.0
uvicorn==0.24.0
gunicorn==21.2.0
APScheduler==3.10.4
requests==2.31.0
pandas==2.1.4
plotly==5.17.0
python-dotenv==1.0.0
Pillow==10.1.0
streamlit-option-menu==0.3.6